refactor: merge filters/ into collections/
The 3 filter components (sidebar, skeleton, sheet) are tightly coupled
to the collections system and shared between collections and search.
Consolidate them into collections/ and rename to drop the "collection-"
prefix on the moved files.

- filter-sidebar.tsx (was collection-filter-sidebar.tsx)
- filter-sidebar-skeleton.tsx (was collection-filter-sidebar-skeleton.tsx)
- filter-sidebar-sheet.tsx (unchanged name)

Removes the separate filters/ directory entirely.
