+++ apps/template/components/collections/filter-sidebar-skeleton.tsx
